chore: renaming and seperate comps
