+++ .eslintrc.json
@@ -1,6 +1,7 @@
 {
   "extends": "next/core-web-vitals",
   "rules": {
-    "no-unused-vars": "warn"
+    "no-unused-vars": "warn",
+    "@next/next/no-img-element": "off"
   }
 }
