+++ app/app/expenses/page.tsx
@@ -3,8 +3,9 @@
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { LuPlus } from 'react-icons/lu'
-import { ExpenseCard, ExpenseWrapper } from './_components/ExpenseGrid'
 import useExpense from './_hooks/useExpense'
+import { ExpenseWrapper } from './_components/ExpenseWrapper'
+import ExpenseCard from './_components/ExpenseCard'
 
 export default function Expenses() {
   const { expenseData, isPending } = useExpense()
